sending srs_toa_ns to T tracer

+++ openair1/PHY/NR_ESTIMATION/nr_ul_estimation.h
@@ -70,7 +70,7 @@ int32_t nr_est_toa_ns_srs(NR_DL_FRAME_PARMS *frame_parms,
                           uint8_t N_arx,
                           uint8_t N_ap,
                           int32_t srs_estimated_channel_freq[N_arx][N_ap][frame_parms->ofdm_symbol_size],
-                          int16_t *srs_toa_ns);
+                          int32_t *srs_toa_ns);
 
 void nr_pusch_ptrs_processing(PHY_VARS_gNB *gNB,
                               NR_DL_FRAME_PARMS *frame_parms,
